Code Quality: Created PreviewPopupService
